+++ src/er_c.rs
@@ -14,6 +14,7 @@ use crate::traits::*;
 use crate::iter::{INContainedIterMut, NContainedIterMut, ContainedIterMut};
 use crate::graph::NodeContainer;
 use std::borrow::Borrow;
+use std::convert::AsRef;
 
 #[cfg(feature = "serde_support")]
 use serde::{Serialize, Deserialize};
@@ -75,6 +76,16 @@ where T: Node,
     prob: f64,
     c_target: f64,
     rng: R,
+}
+
+impl<T, R> AsRef<Graph<T>> for ErEnsembleC<T, R>
+where T: Node,
+      R: rand::Rng
+{
+    #[inline]
+    fn as_ref(&self) -> &Graph<T>{
+        &self.graph
+    }
 }
 
 impl<T, R> Borrow<Graph<T>> for ErEnsembleC<T, R>